Update README.md
## Movie-data-project

#### About the dataset
The dataset contains information about over 10,000 movies ranging from year 
1960 to 2015.

#### Objective of project
The objective of the project is to describe the trends of some
the dataset features, and also examine the relationship between
the revenue feature of the dataset and other features such as budget,
vote count, release year, popularity and runtime using correlation matrix.

#### Packages installed
- Pandas 
- Numpy
- Matplotlib.pyplot
- Seaborn

#### Results
The project uncovered some trends in the datasets features.
The year with highest number of movies released is 2014, 
the genre with the highest vote count is Comedy. 
Steven Spielberg is the director with- highest number of movie titles released,
he has 28 movie titles. We also uncovered that Comedy is the genre which
recorded the highest revenue generated, comedy recorded USD 19,517,216,485.
Furthermore, the movie title with the highest revenue generated is Avatar 
with USD 2,781,505,847.

This project also examined and established that there is a 
positive relationship between revenue and some features such as budget,
vote count and popularity.

Further studies can be done to ascertain the effect of each features
on the revenue generated.